requirements.txt
admin-rights==0.0.1
alembic==1.4.3
amqp==5.0.2
aniso8601==4.1.0
apipkg==1.5
arrotechtools==0.0.2
astroid==2.4.2
attrs==19.3.0
Authlib==0.15.3
autopep8==1.5.4
bcrypt==3.2.0
billiard==3.6.3.0
blinker==1.4
cachetools==4.2.0
celery==5.0.2
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
circuitbreaker==1.3.1
click==6.7
click-didyoumean==0.0.3
click-repl==0.1.6
codecov==2.1.7
coverage==5.2
coveralls==2.1.1
cryptography==3.3.1
dnspython==2.0.0
docopt==0.6.2
email-validator==1.1.2
flake8==3.8.4
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.9.0
Flask-Cors==3.0.8
Flask-DebugToolbar==0.11.0
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-OAuth @ git+https://github.com/mitsuhiko/flask-oauth@af6432621a9add5e55b0d7638e1380f66f3d4ed0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
Flask-WTF==0.14.3
gem==0.1.12
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
greenlet==1.0.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==2.6.0
kombu==5.0.2
lazy-object-proxy==1.4.3
logzero==1.6.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.10.0
mccabe==0.6.1
more-itertools==8.4.0
nose==1.3.7
oauth2==1.9.0.post1
oauthlib==3.1.0
packaging==20.4
paramiko==2.1.1
pluggy==0.13.1
prompt-toolkit==3.0.8
proofdock-chaos-middleware-flask==0.1.3
proofdock-chaos-middleware-python==0.2.3
protobuf==3.14.0
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
requests==2.25.1
rq==1.7.0
rsa==4.7
six==1.15.0
SQLAlchemy==1.1.5
toml==0.10.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==0.16.0
wrapt==1.11.1
WTForms==2.3.3
